Dataset_left_join_is_done
Finally, alhamdulillah I finished the left join for 1000 patients.

by using dask and join file by file.
First, I tried with a CSV file for 1000 patients, and there was a logical error(the sample was reduced to less than 1000 patients). Also, the left join stopped in a very large CSV file. When the final file got to 21 gigas, it just stopped. There were a lot of problems, so I just made the joining file a parquet so it saved memory and did the left join very fast, and the sample remained at 1000 patients.